Communication of one microservice with another with httpclient Done with Authorize

+++ OrderService/BackgroundServices/OutBoxBackground.cs
@@ -37,35 +37,41 @@ public class OutBoxBackground : BackgroundService
 				var genericRepo = scope.ServiceProvider.GetRequiredService<IGenericRepository<AppDbContext, OutBox>>();
 				var unitOfWork = scope.ServiceProvider.GetRequiredService<IUnitOfWork>();
 
+
 				var OutboxOrders = dbContext.OutBoxes.Where(x => x.IsSend == false);
 
-				foreach (var order in OutboxOrders)
+				if (OutboxOrders.Count() > 0)
 				{
-					if (order.IsDelete)
+
+					foreach (var order in OutboxOrders)
 					{
+						if (order.IsDelete)
+						{
+							_rabbitMQPublisher.Publish(order, OutBoxDirect.ExchangeName, OutBoxDirect.QueueName, OutBoxDirect.RoutingWaterMark);
+							_logger.LogInformation($"OutBox sent to RabbitMQ --> {order.Name}");
+
+							genericRepo.Remove(order);
+							_logger.LogInformation($"Outbox order successfully removed --> {order.Name}");
+						}
+						else
+						{
+							order.IsSend = true;
+							genericRepo.UpdateAsync(order);
+							_logger.LogInformation($"Outbox order successfully updated --> {order.Name}");
+						}
+
+						await unitOfWork.CommitAsync();
+
 						_rabbitMQPublisher.Publish(order, OutBoxDirect.ExchangeName, OutBoxDirect.QueueName, OutBoxDirect.RoutingWaterMark);
 						_logger.LogInformation($"OutBox sent to RabbitMQ --> {order.Name}");
-
-						genericRepo.Remove(order);
-						_logger.LogInformation($"Outbox order successfully removed --> {order.Name}");
 					}
-					else
-					{
-						order.IsSend = true;
-						genericRepo.UpdateAsync(order);
-						_logger.LogInformation($"Outbox order successfully updated --> {order.Name}");
-					}
-
-					await unitOfWork.CommitAsync();
-
-					_rabbitMQPublisher.Publish(order, OutBoxDirect.ExchangeName, OutBoxDirect.QueueName, OutBoxDirect.RoutingWaterMark);
-					_logger.LogInformation($"OutBox sent to RabbitMQ --> {order.Name}");
 				}
 				await Task.Delay(TimeSpan.FromSeconds(3), stoppingToken);
 			}
 			catch (Exception ex)
 			{
 				_logger.LogError($"An error occurred: {ex.Message}");
+				await Task.Delay(TimeSpan.FromSeconds(5), stoppingToken);
 			}
 		}
 	}
